added conftest.py
began moving reusable code to it
